fix: add id

+++ src/main/java/com/example/barbershop/dto/ClienteDTO.java
@@ -1,5 +1,5 @@
 package com.example.barbershop.dto;
 
-public record ClienteDTO(Long id, String nome, String telefone, String email, String cpf) {
+public record ClienteDTO(int id, String nome, String telefone, String email, String cpf) {
 
 }
